FEAT.BALANCE,WITHDRAW,ORDERS: добавить Balance/Withdraw endpoints + репозиторий для баланса/выводов + унифицировать user_id как int32
Context:
- Расширяем API авторизованного пользователя: баланс, вывод средств, список заказов/выводов.
- Приводим тип user_id к единому int32 по всему контуру (claims/handler/service/repo).
- Подключаем новые зависимости через ServiceDeps и HandlerDeps.

Changes:
- cmd: прокинуты BalanceDb/WithdrawerDb в ServiceDeps; HandlerDeps заполнен сервисом (Registrar/Auther/TokenChecker/Ordered/Balancer/Withdrawer); добавлен Start/Stop воркера
- handler: исправлен путь withdraw на "/api/user/withdraw"
- handler: UserIdFromCtx возвращает int32; Claims.UserID переведен на int32
- handler: WithdrawHandler — более строгая проверка Content-Type, json.Decoder вместо ReadAll/Unmarshal; уточнена мапа HTTP статусов и логов
- repo/sqlc: добавлен GetOrdersByUserID; добавлен набор запросов withdraws.sql (ensure/lock/withdraw/add/list/balance)
- repo: реализованы GetOrders, Withdraw, ListWithdraws, Balance; обновлены сигнатуры AddUser/GetUserByID/CreateOrder под int32
- service: добавлен balance.go (BalanceDb + метод Balance), расширен withdraws.go (round суммы, ListWithdraws)
- service: Claims.UserID и JwtToken переведены на int32

Implementation details:
- Денежные значения храним как int32 в копейках; наружу — float64 (деление на 100), для withdraw используем math.Round(sum*100)
- GetOrdersByUserID сортирует заказы по uploaded_at DESC
- Withdraw операции выполняются в транзакции: ensure balance row → withdraw-if-enough → add withdrawal

Tests:
- not added

+++ internal/handler/middleware/mw_check_cookie.go
@@ -21,7 +21,7 @@ type TokenChecker interface {
 }
 
 type Claims struct {
-	UserID uint64 `json:"user_id"`
+	UserID int32 `json:"user_id"`
 }
 
 func CheckCookie(cht TokenChecker) func(http.Handler) http.Handler {
